go-kosu: Add `abci` subcommmand
At this point it's still tricky to implement the --validator flag here.
This will create a circular dependency between the Witness and the ABCI
server, the witness needs the ABCI server to be started to query rounds, which is
implemented by tendermint and not go-kosu anymore.

There are possible solutions to this problem that I'd explore later.

For now, this is working, and you can alternate, running as a
stand-alone application and then switch to ABCI mode using the same
homedir.